Fix code quality issues reported by CodeScene
- Refactor find_sessions_to_compress to reduce complexity by extracting
  _session_needs_compression helper function
- Remove code duplication across test files by moving shared fixtures
  (temp_sessions_dir, create_wav_file) to tests/conftest.py
